Swapped JExifTool for mjeanroy/exiftool
