--- converter/repo/be/pw/jexif/0.0.9/jexif-0.0.9.jar.sha1
@@ -1,1 +1,0 @@
-4f0f66eeaa09ef1b0674a6310484159348696812
