--- converter/repo/be/pw/jexif/0.0.9/jexif-0.0.9.pom
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>be.pw</groupId>
-  <artifactId>jexif</artifactId>
-  <version>0.0.9</version>
-</project>
