--- converter/repo/be/pw/jexif/maven-metadata.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<metadata>
-  <groupId>be.pw</groupId>
-  <artifactId>jexif</artifactId>
-  <versioning>
-    <release>0.0.9</release>
-    <versions>
-      <version>0.0.9</version>
-    </versions>
-    <lastUpdated>20171005070313</lastUpdated>
-  </versioning>
-</metadata>
