--- converter/repo/be/pw/jexif/maven-metadata.xml.sha1
@@ -1,1 +1,0 @@
-fc4c42771f9e09cb9f8921390a4220682f8dd915
